app/server/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
Werkzeug==3.0.1
python-dotenv==1.0.0
Cerberus==1.3.5
numpy==1.26.2
pandas==2.1.3
seaborn==0.13.0
matplotlib==3.8.2
plotly==5.18.0
pickle5==0.0.11
scikit-learn==1.3.2

